move non-oauthserver packages out
