+++ pkg/apiserver/authentication/internaloauth/expirationvalidator.go
@@ -1,4 +1,4 @@
-package registry
+package internaloauth
 
 import (
 	"errors"
@@ -6,13 +6,12 @@ import (
 
 	userv1 "github.com/openshift/api/user/v1"
 	"github.com/openshift/origin/pkg/oauth/apis/oauth"
-	"github.com/openshift/origin/pkg/oauthserver/authenticator"
 )
 
 var errExpired = errors.New("token is expired")
 
-func NewExpirationValidator() authenticator.OAuthTokenValidator {
-	return authenticator.OAuthTokenValidatorFunc(
+func NewExpirationValidator() OAuthTokenValidator {
+	return OAuthTokenValidatorFunc(
 		func(token *oauth.OAuthAccessToken, _ *userv1.User) error {
 			if token.ExpiresIn > 0 {
 				if expire(token).Before(time.Now()) {
