+++ pkg/apiserver/authentication/internaloauth/tokenauthenticator.go
@@ -1,4 +1,4 @@
-package registry
+package internaloauth
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -8,23 +8,21 @@ import (
 	userclient "github.com/openshift/client-go/user/clientset/versioned/typed/user/v1"
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
 	oauthclient "github.com/openshift/origin/pkg/oauth/generated/internalclientset/typed/oauth/internalversion"
-	"github.com/openshift/origin/pkg/oauthserver/authenticator"
-	"github.com/openshift/origin/pkg/oauthserver/userregistry/identitymapper"
 )
 
 type tokenAuthenticator struct {
 	tokens      oauthclient.OAuthAccessTokenInterface
 	users       userclient.UserInterface
-	groupMapper identitymapper.UserToGroupMapper
-	validators  authenticator.OAuthTokenValidator
+	groupMapper UserToGroupMapper
+	validators  OAuthTokenValidator
 }
 
-func NewTokenAuthenticator(tokens oauthclient.OAuthAccessTokenInterface, users userclient.UserInterface, groupMapper identitymapper.UserToGroupMapper, validators ...authenticator.OAuthTokenValidator) kauthenticator.Token {
+func NewTokenAuthenticator(tokens oauthclient.OAuthAccessTokenInterface, users userclient.UserInterface, groupMapper UserToGroupMapper, validators ...OAuthTokenValidator) kauthenticator.Token {
 	return &tokenAuthenticator{
 		tokens:      tokens,
 		users:       users,
 		groupMapper: groupMapper,
-		validators:  authenticator.OAuthTokenValidators(validators),
+		validators:  OAuthTokenValidators(validators),
 	}
 }
