+++ pkg/apiserver/authentication/internaloauth/uidvalidator.go
@@ -1,17 +1,16 @@
-package registry
+package internaloauth
 
 import (
 	"fmt"
 
-	"github.com/openshift/origin/pkg/oauth/apis/oauth"
-	"github.com/openshift/origin/pkg/oauthserver/authenticator"
 	userapi "github.com/openshift/api/user/v1"
+	"github.com/openshift/origin/pkg/oauth/apis/oauth"
 )
 
 const errInvalidUIDStr = "user.UID (%s) does not match token.userUID (%s)"
 
-func NewUIDValidator() authenticator.OAuthTokenValidator {
-	return authenticator.OAuthTokenValidatorFunc(
+func NewUIDValidator() OAuthTokenValidator {
+	return OAuthTokenValidatorFunc(
 		func(token *oauth.OAuthAccessToken, user *userapi.User) error {
 			if string(user.UID) != token.UserUID {
 				return fmt.Errorf(errInvalidUIDStr, user.UID, token.UserUID)
