+++ pkg/oauthserver/authenticator/interfaces.go
@@ -3,9 +3,7 @@ package authenticator
 import (
 	"k8s.io/apiserver/pkg/authentication/user"
 
-	"github.com/openshift/origin/pkg/oauth/apis/oauth"
 	"github.com/openshift/origin/pkg/oauthserver/api"
-	userapi "github.com/openshift/api/user/v1"
 )
 
 type Assertion interface {
@@ -14,29 +12,4 @@ type Assertion interface {
 
 type Client interface {
 	AuthenticateClient(client api.Client) (user.Info, bool, error)
-}
-
-type OAuthTokenValidator interface {
-	Validate(token *oauth.OAuthAccessToken, user *userapi.User) error
-}
-
-var _ OAuthTokenValidator = OAuthTokenValidatorFunc(nil)
-
-type OAuthTokenValidatorFunc func(token *oauth.OAuthAccessToken, user *userapi.User) error
-
-func (f OAuthTokenValidatorFunc) Validate(token *oauth.OAuthAccessToken, user *userapi.User) error {
-	return f(token, user)
-}
-
-var _ OAuthTokenValidator = OAuthTokenValidators(nil)
-
-type OAuthTokenValidators []OAuthTokenValidator
-
-func (v OAuthTokenValidators) Validate(token *oauth.OAuthAccessToken, user *userapi.User) error {
-	for _, validator := range v {
-		if err := validator.Validate(token, user); err != nil {
-			return err
-		}
-	}
-	return nil
 }
